+++ backend/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "express-project-starter",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "start": "per-env",
+    "start:development": "nodemon -r dotenv/config ./bin/www",
+    "start:production": "node ./bin/www"
+  },
+  "dependencies": {
+    "bcrypt": "^5.1.0",
+    "connect-session-sequelize": "^7.0.4",
+    "cookie-parser": "~1.4.4",
+    "cors": "^2.8.5",
+    "debug": "~2.6.9",
+    "express": "~4.16.1",
+    "express-session": "^1.17.1",
+    "http-errors": "~1.6.3",
+    "jsonwebtoken": "^9.0.0",
+    "morgan": "~1.9.1",
+    "node-fetch": "^2.6.11",
+    "per-env": "^1.0.2",
+    "pg": "^8.4.2",
+    "pug": "2.0.4",
+    "sequelize": "^6.31.1",
+    "sequelize-cli": "^5.5.1"
+  },
+  "devDependencies": {
+    "dotenv": "^8.2.0",
+    "dotenv-cli": "^4.0.0",
+    "nodemon": "^2.0.6"
+  }
+}
