--- package.json
@@ -1,70 +1,0 @@
-{
-  "name": "roundtrip",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "start": "per-env",
-    "start:development": "nodemon -r dotenv/config ./bin/www",
-    "start:production": "node ./bin/www",
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@emotion/react": "^11.11.0",
-    "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.13.2",
-    "@mui/x-date-pickers": "^6.9.0",
-    "@react-google-maps/api": "^2.18.1",
-    "@reduxjs/toolkit": "^1.9.5",
-    "@types/google.maps": "^3.53.2",
-    "@types/googlemaps": "^3.43.3",
-    "bcrypt": "^5.1.0",
-    "connect-session-sequelize": "^7.0.4",
-    "cookie-parser": "~1.4.4",
-    "cors": "^2.8.5",
-    "dayjs": "^1.11.8",
-    "debug": "~2.6.9",
-    "express": "4.18.2",
-    "express-session": "^1.17.1",
-    "hamburger-react": "^2.5.0",
-    "http-errors": "~1.6.3",
-    "jsonwebtoken": "^9.0.0",
-    "morgan": "~1.9.1",
-    "node-fetch": "^2.6.11",
-    "per-env": "^1.0.2",
-    "pg": "^8.4.2",
-    "pug": "3.0.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-material-ui-carousel": "^3.4.2",
-    "react-redux": "^8.0.5",
-    "redux-logger": "^3.0.6",
-    "sass": "^1.62.1",
-    "sequelize": "^6.31.1",
-    "sequelize-cli": "^5.5.1"
-  },
-  "devDependencies": {
-    "@types/node": "^20.2.5",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.57.1",
-    "@typescript-eslint/parser": "^5.57.1",
-    "@vitejs/plugin-react": "^4.0.0",
-    "@vitejs/plugin-react-refresh": "^1.3.6",
-    "autoprefixer": "^10.4.14",
-    "dotenv": "^16.1.3",
-    "dotenv-cli": "^4.0.0",
-    "eslint": "^8.38.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.3.4",
-    "nodemon": "^2.0.6",
-    "postcss": "^8.4.23",
-    "react-router-dom": "^6.11.2",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.0.2",
-    "vite": "^4.3.2"
-  }
-}
